pyproject.toml
[project]
name = "mcp-pagoda"
version = "0.0.1"
description = "A MCP Server for Pagoda"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["pagoda", "mcp"]
license = { text = "MIT" }
dependencies = [
    "requests>=2.32.3",
    "click>=8.1.8",
    "mcp>=1.10.0",
    "anyio>=4.5",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.5.2",
    "sse-starlette>=1.6.1",
    "uvicorn>=0.23.1; sys_platform != 'emscripten'",
    "python-dotenv>=1.1.1",
]

[project.scripts]
mcp-server = "mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.399",
    "ruff>=0.11.5",
    "pytest>=8.3.5",
    "types-requests>=2.32.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
